first working version half dnd inside constructor
