+++ src/components/ConstructorElementWrap/ConstructorElementWrap.module.css
@@ -52,5 +52,3 @@
   align-self: center;
   margin-left: 22px;
 }
-
-
